feat: Implement snowflake ID generation for models - Add snowflake library for unique ID generation - Implement snowflake ID generation for Session, VisitRecord, and GeoLocation models - Initialize snowflake node in utils package

+++ models/user-analytics.go
@@ -3,7 +3,7 @@ package models
 import "time"
 
 type Session struct {
-	ID         uint      `json:"id" gorm:"primaryKey"`
+	ID         int64     `json:"id" gorm:"primaryKey"`
 	FirstSeen  time.Time `json:"first_seen"`
 	LastSeen   time.Time `json:"last_seen"`
 	VisitCount int       `json:"visit_count"`
@@ -16,10 +16,10 @@ type Session struct {
 }
 
 type VisitRecord struct {
-	ID            uint         `json:"id" gorm:"primaryKey"`
+	ID            int64        `json:"id" gorm:"primaryKey"`
 	GeoLocation   *GeoLocation `json:"geo_location" gorm:"foreignKey:GeoLocationId"` // Associated GeoLocation
-	GeoLocationId *uint        `json:"geo_location_id"`                              // Foreign key to GeoLocation
-	SessionID     uint         `json:"session_id"`                                   // Unique session identifier
+	GeoLocationId *int64       `json:"geo_location_id"`                              // Foreign key to GeoLocation
+	SessionID     int64        `json:"session_id"`                                   // Unique session identifier
 	ShortURLID    string       `json:"short_url_id"`                                 // Unique short URL identifier
 	VisitTime     time.Time    `json:"visit_time"`                                   // Time of visit
 	CreatedAt     time.Time    `json:"create_at"`
@@ -27,7 +27,7 @@ type VisitRecord struct {
 }
 
 type GeoLocation struct {
-	ID          uint    `json:"id" gorm:"primaryKey"`
+	ID          int64   `json:"id" gorm:"primaryKey"`
 	Ip          string  `json:"ip"`
 	Country     string  `json:"country"`
 	CountryCode string  `json:"countryCode"`
